perf: lazily parse attributes & other improvments
This is a combined commit containing lots of little optimizations. The result
is a 25-50% speed up. There is still a lot
of room for improvement though.

```
tl                 : 0.999355 ms/file ± 0.736266
htmlparser2-new    : 1.45273 ms/file ± 2.20925
htmlparser2        : 2.17451 ms/file ± 2.98730
```

This doesn't break the external API except removing the `onattribute`
event which doesn't make sense in this model. Other than that everything
should work as expected.